+++ tests/testthat/_snaps/rix/bioc-devel_default.nix
@@ -1,0 +1,30 @@
+
+let
+  pkgs = import (fetchTarball "https://github.com/rstats-on-nix/nixpkgs/archive/refs/heads/r-bioc-devel.tar.gz") {};
+ 
+  rpkgs = builtins.attrValues {
+    inherit (pkgs.rPackages) 
+      S4Vectors;
+  };
+    
+  system_packages = builtins.attrValues {
+    inherit (pkgs) 
+      R
+      glibcLocales
+      nix;
+  };
+  
+in
+
+pkgs.mkShell {
+  LOCALE_ARCHIVE = if pkgs.system == "x86_64-linux" then "${pkgs.glibcLocales}/lib/locale/locale-archive" else "";
+  LANG = "en_US.UTF-8";
+   LC_ALL = "en_US.UTF-8";
+   LC_TIME = "en_US.UTF-8";
+   LC_MONETARY = "en_US.UTF-8";
+   LC_PAPER = "en_US.UTF-8";
+   LC_MEASUREMENT = "en_US.UTF-8";
+
+  buildInputs = [  rpkgs  system_packages   ];
+  
+}